pyproject.toml
[project]
name = "real-voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi==0.115.8",
    "httpx==0.28.1",
    "ipython==8.13.2",
    "langchain-community>=0.3.27",
    "livekit-agents[cartesia,deepgram,elevenlabs,google,openai,silero,turn-detector]~=1.0",
    "livekit-plugins-cartesia>=1.0.11",
    "livekit-plugins-deepgram>=1.0.11",
    "livekit-plugins-google>=1.0.11",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.0.11",
    "livekit-plugins-silero>=1.0.11",
    "mem0ai>=0.1.115",
    "python-dotenv==1.0.1",
    "requests>=2.32.4",
    "uvicorn==0.34.0",
]
